+++ src/app/game/game-mode/mode/promode-mode.ts
@@ -3,31 +3,31 @@ import { CountdownState } from '../base-game-mode.ts/countdown-state';
 import { GameLoopState } from '../base-game-mode.ts/game-loop-state';
 import { GameOverState } from '../base-game-mode.ts/game-over-state';
 import { ResetState } from '../base-game-mode.ts/reset-state';
+import { SetPromodeTempVisionState } from '../promode-game-mode/set-promode-temp-vision-state';
 import { SetupState } from '../base-game-mode.ts/setup-state';
 import { BaseMode } from './base-mode';
 import { BaseState } from '../state/base-state';
 import { StateData } from '../state/state-data';
 import { WaitState } from '../base-game-mode.ts/wait-state';
-import { VisionState } from '../base-game-mode.ts/vision-state';
 
-export class StandardData implements StateData {}
+export class PromodeData implements StateData {}
 
-export class StandardMode extends BaseMode<StandardData> {
+export class PromodeMode extends BaseMode<PromodeData> {
 	protected setupStates() {
 		return [
 			new SetupState(),
 			new WaitState(2),
 			new CityDistributeState(),
-			new VisionState(),
+			new SetPromodeTempVisionState(),
 			new WaitState(2),
 			new CountdownState(),
 			new GameLoopState(),
 			new GameOverState(),
 			new ResetState(),
-		] as BaseState<StandardData>[];
+		] as BaseState<PromodeData>[];
 	}
 
-	protected setupData(): StandardData {
-		return new StandardData();
+	protected setupData(): PromodeData {
+		return new PromodeData();
 	}
 }
